refactor(DatePicker): Abstract to extra hook
Parsing on input keys can move to a separate hook
to address code complexity.
